+++ src-tauri/session-service/ios/Package.swift
@@ -1,8 +1,8 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.5
 import PackageDescription
 let package = Package(
     name: "tauri-plugin-session-service",
-    platforms: [.iOS(.v13)],
+    platforms: [.iOS(.v15)],
     products: [.library(name: "tauri-plugin-session-service", type: .static, targets: ["tauri-plugin-session-service"])],
     dependencies: [.package(name: "Tauri", path: "../.tauri/tauri-api")],
     targets: [.target(name: "tauri-plugin-session-service", dependencies: [.byName(name: "Tauri")], path: "Sources")]